MM_30482 : Removed memberCount and timezoneCount from GroupsInTeam
