+++ app/database/admin/data_operator/utils/test.ts
@@ -2,11 +2,12 @@
 // See LICENSE.txt for license information.
 
 import DataOperator from '@database/admin/data_operator';
+import {createPostsChain, sanitizePosts} from '@database/admin/data_operator/utils/post';
+import {sanitizeReactions} from '@database/admin/data_operator/utils/reaction';
 import DatabaseManager from '@database/admin/database_manager';
 import {DatabaseType} from '@typings/database/enums';
 import {RawPost} from '@typings/database/database';
 
-import {createPostsChain, sanitizePosts, sanitizeReactions} from './index';
 import {mockedPosts, mockedReactions} from './mock';
 
 jest.mock('@database/admin/database_manager');
