+++ app/database/server/models/groups_in_team.ts
@@ -31,14 +31,8 @@ export default class GroupsInTeam extends Model {
     /** group_id : The foreign key to the related Group record */
     @field('group_id') groupId!: string;
 
-    /** member_count : The number of users in that group */
-    @field('member_count') memberCount!: number;
-
     /** team_id : The foreign key to the related Team record */
     @field('team_id') teamId!: string;
-
-    /** timezone_count : The number of timezones */
-    @field('timezone_count') timezoneCount!: number;
 
     /** team : The related record to the parent Team model */
     @immutableRelation(TEAM, 'team_id') team!: Relation<Team>;
